requirements.txt
matplotlib==3.8.4
numpy==2.1.1
opencv_python==4.9.0.80
opencv_python_headless==4.10.0.84
pyzmq==26.2.0
scikit_learn==1.5.2
ultralytics==8.2.12
